move licenser in libbeat
