+++ x-pack/functionbeat/core/license_client.go
@@ -10,7 +10,7 @@ import (
 	"github.com/elastic/beats/libbeat/beat"
 	"github.com/elastic/beats/libbeat/common/atomic"
 	"github.com/elastic/beats/libbeat/logp"
-	"github.com/elastic/beats/x-pack/functionbeat/licenser"
+	"github.com/elastic/beats/x-pack/libbeat/licenser"
 )
 
 var errInvalidLicense = errors.New("invalid license detected, cannot publish events")
