+++ x-pack/libbeat/licenser/check.go
