+++ x-pack/libbeat/licenser/check_test.go
