+++ x-pack/libbeat/licenser/license_test.go
