+++ x-pack/libbeat/licenser/licensetype_string.go
