+++ x-pack/libbeat/licenser/manager.go
