+++ x-pack/libbeat/licenser/manager_test.go
