+++ x-pack/libbeat/licenser/testdata/x-pack-trial-6.4.0.json
